weightpred/requirements.txt
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-crispy-forms==1.8.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
joblib==0.14.1
numpy==1.16.4
pandas==0.24.2
Pillow==6.2.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
